requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pandas==2.1.3
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1